+++ index.test.js
@@ -1318,6 +1318,46 @@ describe("serverless-wsgi", () => {
         sandbox.restore();
       });
     });
+
+    it("loads wsgi app from individually packed module", () => {
+      var plugin = new Plugin(
+        {
+          config: { servicePath: "/tmp" },
+          service: {
+            provider: { runtime: "python2.7" },
+            package: { individually: true },
+            custom: { wsgi: { app: "site/api.app" } },
+            functions: {
+              api: { handler: "wsgi.handler", module: "site" }
+            }
+          },
+          classes: { Error: Error },
+          cli: { log: () => {} }
+        },
+        {}
+      );
+
+      var sandbox = sinon.createSandbox();
+      var hasbinStub = sandbox.stub(hasbin, "sync").returns(true);
+      var procStub = sandbox.stub(child_process, "spawnSync").returns({});
+      plugin.hooks["wsgi:serve:serve"]().then(() => {
+        expect(hasbinStub.calledWith("python2.7")).to.be.true;
+        expect(
+          procStub.calledWith(
+            "python2.7",
+            [
+              path.resolve(__dirname, "serve.py"),
+              "/tmp/site",
+              "api.app",
+              5000,
+              "localhost"
+            ],
+            { stdio: "inherit" }
+          )
+        ).to.be.true;
+        sandbox.restore();
+      });
+    });
   });
 
   describe("install", () => {